requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
tabulate>=0.9.0
requests>=2.28.0
gunicorn>=21.0.0
flasgger==0.9.7.1
